requirements.txt
Django==3.2.8
psycopg2==2.9.1
Pillow==8.3.2
pycryptodome==3.11.0
